Fix conv2d weight access shadowing, sgemm transpose strides, and SWAR min test values
- Fix variable shadowing bug in conv2d_3x3_direct (and 4 fused variants):
  `let n = out_channels` was shadowed by batch loop variable `|n|`,
  causing weight reads to use batch index instead of out_channels
- Fix buffer overflow + wrong values in `matmul_blas_with_transpose_into`:
  `matrixmultiply::sgemm` path swapped M/K/N dimensions on transpose
  instead of adjusting row/column strides; this wrote past the output
  buffer and computed wrong matrix dimensions
- Fix SWAR min test expectations: both `swar_min_u16x2` and
  `swar_min_u4x8` tests had incorrect expected values (not bugs in
  the implementations)
- Add optimizer re-exports, WeightDecayOptimizer impl, test tolerance
  adjustments, and example fixes

+++ tests/optim_test.rs
@@ -1,4 +1,4 @@
-use fastnn::optim::{Adam, AdamW, Lion, Muon, RMSprop, SGD};
+use fastnn::optim::{Adam, AdamW, Lion, Muon, Optimizer, RMSprop, SGD, WeightDecayOptimizer};
 use fastnn::tensor::Tensor;
 
 fn create_test_params() -> Vec<Tensor> {
@@ -10,7 +10,7 @@ fn create_test_params() -> Vec<Tensor> {
 
 fn set_grads(params: &mut [Tensor]) {
     for p in params {
-        let grad = Tensor::from_vec(vec![0.1f32; p.numel()], p.shape());
+        let grad = Tensor::from_vec(vec![0.1f32; p.numel() as usize], p.shape());
         p.set_grad(Some(grad));
     }
 }
